+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Fixed bug in `db.Topic()` with send response to stop partition message
+
 ## v3.32.1
 * Fixed flaky TestTable
 * Renamed topic events in `trace.Details` enum